requirements.txt
flask
PyMySql
setuptools
oslo.db
oslo.messaging
oslo.config
oslo.concurrency
pymysql
pydantic
taskflow
cotyledon
fastapi 
sqlmodel 
python-dotenv
celery
flower